Implement folder selection in open dialogs for GTK, Gtk3, and WinUI backends
Add OpenDialogOptions.singleKindSelectionMode to express whether an open
dialog should select files or directories on backends that only support
one kind of item per dialog. Files are preferred when both are allowed,
preserving the existing behavior of the GTK and WinUI backends.

- GtkBackend/Gtk3Backend: use FileChooserAction.selectFolder instead of
  .open when the dialog is directories-only
- WinUIBackend: split showOpenDialog into showFileOpenDialog and a new
  FolderPicker-based showFolderOpenDialog for directories-only dialogs;
  selecting multiple folders is unsupported by FolderPicker and guarded
  by a precondition
- Add OpenDialogOptionsTests covering the selection mode logic

Also fix the Gtk3 module build on Windows:
- Use gulong directly for signal handler IDs instead of converting
  through UInt
- Rename Gtk3CHelpers' max() to static scui_max() to avoid symbol
  clashes

Note: Used Claude Code as LLM Coding support

+++ Sources/Gtk3/GObject.swift
@@ -33,7 +33,7 @@ open class GObject: GObjectRepresentable {
         g_object_unref(gobjectPointer)
     }
 
-    private var signals: [(UInt, Any)] = []
+    private var signals: [(gulong, Any)] = []
 
     open func registerSignals() {}
